+++ packages/config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @verdaccio/config
 
+## 7.0.0-next-7.17
+
+### Patch Changes
+
+- 6e764e3: feat: add support for npm owner
+- Updated dependencies [6e764e3]
+  - @verdaccio/core@7.0.0-next-7.17
+  - @verdaccio/utils@7.0.0-next-7.17
+
 ## 7.0.0-next-7.16
 
 ### Patch Changes