+++ packages/plugins/htpasswd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 12.0.0-next-7.17
+
+### Patch Changes
+
+- Updated dependencies [6e764e3]
+  - @verdaccio/core@7.0.0-next-7.17
+  - @verdaccio/file-locking@12.0.0-next.1
+
 ## 12.0.0-next-7.16
 
 ### Patch Changes